one-bangkok-app-trunk/automate/backend/requirements.txt
robotframework==4.1.2
robotframework-Requests==0.9.2
pyyaml==5.3.1
robotframework-databaselibrary==1.2.4
pymysql==1.0.2
git+https://github.com/peterservice-rnd/robotframework-rabbitmq.git
pycryptodome===3.11.0
pika==0.12.0
ulid-py
psycopg2-binary